NOJIRA sanity checks caught the public
