+++ web/src/main/java/no/nav/foreldrepenger/web/app/tjenester/fagsak/FagsakRestTjeneste.java
@@ -35,15 +35,9 @@ import no.nav.foreldrepenger.behandlingslager.aktør.PersoninfoBasis;
 import no.nav.foreldrepenger.behandlingslager.fagsak.Dekningsgrad;
 import no.nav.foreldrepenger.behandlingslager.fagsak.Fagsak;
 import no.nav.foreldrepenger.domene.typer.Saksnummer;
-import no.nav.foreldrepenger.web.app.rest.ResourceLink;
-import no.nav.foreldrepenger.web.app.tjenester.aktoer.AktoerIdDto;
-import no.nav.foreldrepenger.web.app.tjenester.aktoer.AktoerRestTjeneste;
-import no.nav.foreldrepenger.web.app.tjenester.behandling.BehandlingRestTjeneste;
 import no.nav.foreldrepenger.web.app.tjenester.behandling.dto.AsyncPollingStatus;
 import no.nav.foreldrepenger.web.app.tjenester.behandling.dto.Redirect;
 import no.nav.foreldrepenger.web.app.tjenester.behandling.dto.behandling.ProsessTaskGruppeIdDto;
-import no.nav.foreldrepenger.web.app.tjenester.behandling.historikk.HistorikkRestTjeneste;
-import no.nav.foreldrepenger.web.app.tjenester.dokument.DokumentRestTjeneste;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.app.FagsakApplikasjonTjeneste;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.app.FagsakSamlingForBruker;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.dto.FagsakBackendDto;
@@ -51,7 +45,6 @@ import no.nav.foreldrepenger.web.app.tjenester.fagsak.dto.FagsakDto;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.dto.PersonDto;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.dto.SaksnummerDto;
 import no.nav.foreldrepenger.web.app.tjenester.fagsak.dto.SokefeltDto;
-import no.nav.foreldrepenger.web.app.util.RestUtils;
 import no.nav.vedtak.sikkerhet.abac.BeskyttetRessurs;
 import no.nav.vedtak.sikkerhet.abac.BeskyttetRessursActionAttributt;
 
@@ -173,26 +166,9 @@ public class FagsakRestTjeneste {
             Integer antallBarn = info.getAntallBarn();
             var dekningsgrad = info.getDekningsgrad().map(d -> d.getVerdi()).orElse(null);
             dtoer.add(new FagsakDto(fagsak, personDto, fødselsdato, antallBarn, kanRevurderingOpprettes, fagsak.getSkalTilInfotrygd(),
-                    fagsak.getRelasjonsRolleType(), dekningsgrad, lagLenker(fagsak)));
+                    fagsak.getRelasjonsRolleType(), dekningsgrad, FagsakApplikasjonTjeneste.lagLenker(fagsak)));
         }
         return dtoer;
-    }
-
-    public static List<ResourceLink> lagLenker(Fagsak fagsak) {
-        List<ResourceLink> lenkene = new ArrayList<>();
-        var saksnummer = new SaksnummerDto(fagsak.getSaksnummer());
-        lenkene.add(get(AktoerRestTjeneste.AKTOER_INFO_PATH, "sak-aktoer-person", new AktoerIdDto(fagsak.getAktørId().getId())));
-        lenkene.add(get(FAGSAK_PATH, "fagsak", saksnummer));
-        lenkene.add(get(BehandlingRestTjeneste.HANDLING_RETTIGHETER_V2_PATH, "handling-rettigheter-v2", saksnummer));
-        lenkene.add(get(HistorikkRestTjeneste.HISTORIKK_PATH, "sak-historikk", saksnummer));
-        lenkene.add(get(DokumentRestTjeneste.DOKUMENTER_PATH, "sak-dokumentliste", saksnummer));
-        lenkene.add(get(BehandlingRestTjeneste.BEHANDLINGER_ALLE_PATH, "sak-alle-behandlinger", saksnummer));
-        lenkene.add(get(BehandlingRestTjeneste.ANNEN_PART_BEHANDLING_PATH, "sak-annen-part-behandling", saksnummer));
-        return lenkene;
-    }
-
-    static ResourceLink get(String path, String rel, Object dto) {
-        return ResourceLink.get(RestUtils.getApiPath(path), rel, dto);
     }
